[refactor/#87] MapSearch Refactoring
