+++ app/src/main/java/com/teamfillin/fillin/presentation/map/search/MapSearchActivity.kt
@@ -1,19 +1,26 @@
-package com.teamfillin.fillin.presentation.map
+package com.teamfillin.fillin.presentation.map.search
 
-import android.app.Activity
 import android.content.Intent
 import android.graphics.Color
 import android.os.Bundle
+import android.util.Log
 import android.view.MenuItem
 import android.view.inputmethod.EditorInfo
 import androidx.activity.viewModels
 import androidx.core.view.isVisible
 import androidx.core.widget.doAfterTextChanged
+import androidx.lifecycle.flowWithLifecycle
+import androidx.lifecycle.lifecycleScope
 import com.teamfillin.fillin.R
 import com.teamfillin.fillin.core.base.BindingActivity
 import com.teamfillin.fillin.core.context.toast
+import com.teamfillin.fillin.core.view.UiState
 import com.teamfillin.fillin.databinding.ActivityMapSearchBinding
+import com.teamfillin.fillin.presentation.map.CustomDecoration
+import com.teamfillin.fillin.presentation.map.SearchListAdapter
 import dagger.hilt.android.AndroidEntryPoint
+import kotlinx.coroutines.flow.launchIn
+import kotlinx.coroutines.flow.onEach
 
 @AndroidEntryPoint
 class MapSearchActivity : BindingActivity<ActivityMapSearchBinding>(R.layout.activity_map_search) {
@@ -24,7 +31,7 @@ class MapSearchActivity : BindingActivity<ActivityMapSearchBinding>(R.layout.act
             putExtra("id", it.id)
             putExtra("name", it.name)
         }
-        setResult(Activity.RESULT_OK, intent)
+        setResult(RESULT_OK, intent)
         finish()
     }
 
@@ -81,20 +88,19 @@ class MapSearchActivity : BindingActivity<ActivityMapSearchBinding>(R.layout.act
     }
 
     private fun observe() {
-        viewModel.resultSearch.observe(this) {
-            when (it) {
-                is MapSearchViewModel.StudioSearchState.Studios -> {
-                    searchListAdapter.submitList(it.studios)
+        viewModel.resultSearch.flowWithLifecycle(lifecycle)
+            .onEach {
+                binding.uiState = it.getUiStateModel()
+                when (it) {
+                    is UiState.Success -> {
+                        searchListAdapter.submitList(it.data)
+                    }
+                    is UiState.Failure -> {
+                        toast(it.msg)
+                    }
+                    else -> {}
                 }
-                is MapSearchViewModel.StudioSearchState.Failure -> {
-                    toast("서버 오류")
-                }
-            }
-            binding.apply {
-                clNoResult.isVisible = it is MapSearchViewModel.StudioSearchState.Empty
-                rvLocationInfo.isVisible = it is MapSearchViewModel.StudioSearchState.Studios
-            }
-        }
+            }.launchIn(lifecycleScope)
     }
 
     private fun editTextBlankCheck() {
